upstream: [media] cx23885[v4]: Fix interrupt storm when enabling IR receiver
Apparently the Flatiron genereates an interrupt after the built-in self
test for each of its left and right channels has completed.

Apparently Conexant wire-OR'ed the Flatiron's interrupt output with the
interrupt output of the CX23885 A/V core.

Those interrupts need to be handled, otherwise, they generate an
interrrupt request storm.

So:

- Add flatiron readreg and writereg functions prototypes
  on a new header file;
- Modify the av interrupt handler to cleanup flatiron IRQs if no other
  interrupt handling happens.

Signed-off-by: Luis Alves <ljalvs@gmail.com>
Acked-by: Andy Walls <awalls@md.metrocast.net>
Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>