vrange: Add support for volatile ranges on file mappings
Like with the mm struct, this patch add basic support for
volatile ranges on file address_space structures. This allows
for volatile ranges to be set on mmapped files that can be
shared between processes.

The semantics on the volatile range sharing is that the
volatility is shared, just as the data is shared. Thus
if one process marks the range as volatile, the data is
volatile in all processes that have those pages mapped.

It is advised that processes coodinate when using volatile
ranges on shared mappings (much as they must coordinate when
writing to shared data).

XXX: Likely squish down file parts of dynamic vroot allocation into
this patch

Signed-off-by: John Stultz <john.stultz@linaro.org>