ARM: dts: exynos4412-slp_pq: always turn on ESAFEOUT1
This patch is only WIP(work-in-progress) patch. Max77693's ESAFEOUT1 is used
for AP-USB VBUS. To enable USB for AP, it should be turned on always or when
a cable is connected. Extcon driver can detect the cable status, so ESAFEOUT1
can be brought up by the extcon. Until supporting it, we should turn ESAFEOUT1
on boot time.

Signed-off-by: Chanho Park <chanho61.park@samsung.com>