usb: gadget: s3c-hsotg: specify device mode before initializing phy
The otg module can be used both host and device mode. According to the boot
sequence, device mode will be initialized first and host mode will be
initialized later. Thus, the phy always remained in the host mode.
To address this problem, we should clear the 'host' value before initializing
the phy driver.

Signed-off-by: Chanho Park <chanho61.park@samsung.com>