sched: replace INIT_COMPLETION with reinit_completion
For the casual device driver writer, it is hard to remember when to use
init_completion (to init a completion structure) or INIT_COMPLETION (to
*reinit* a completion structure).  Furthermore, while all other
completion functions exepct a pointer as a parameter, INIT_COMPLETION
does not.  To make it easier to remember which function to use and to
make code more readable, introduce a new inline function with the proper
name and consistent argument type.  Update the kernel-doc for
init_completion while we are here.

Change-Id: I7bd11b98add6341cccd56b82f25b40d48e0039f0
Signed-off-by: Wolfram Sang <wsa@the-dreams.de>
Acked-by: Linus Walleij <linus.walleij@linaro.org> (personally at LCE13)
Cc: Ingo Molnar <mingo@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Cc:대인기 <inki.dae@samsung.com>
Cc:박찬호 <chanho61.park@samsung.com>
Cc:김승우 <sw0312.kim@samsung.com>
Cc:심준영 <jy0922.shim@samsung.com>
Cc:이동화 <dh09.lee@samsung.com>
Cc:최찬우 <cw00.choi@samsung.com>
Cc:정재훈 <jh80.chung@samsung.com>