Revert "abi-checker - tool directory location update."
This reverts commit 952c6d9596e18f9df43174f513ce84ebe7cd6ca2.
Do we really need this tool? I found below message due to only a commit(f7d3c6)

[  367s] 	Kernel symbols version statistics
[  367s] 	-------------------------------------------------
[  367s] 	Symbols in actual kernel/module . 6901
[  367s] 	Symbols in new kernel/module .... 6900
[  367s] 	New symbols ..................... 0
[  367s] 	Removed symbols ................. 1
[  367s] 	Changed symbols ................. 2
[  367s] 	Unchanged symbols ............... 6898
[  367s]
[  367s]
[  367s] 	Changes !!!

The kernel is always changed if we pulled sources from linus tree or to apply
stable patches. In the case, we have to update the horrible ABI changes.
Worse still, we should write more patches to update the ABI change.
This thing makes it difficult to develop and to apply latest kernel patches.
Why do we need this tool? If the third-party developer wants to know the ABI
changes, it is sufficient for looking the latest sources :)
Why do you open the hell gate only few advantages?

Change-Id: I725c42847254b04ca674364802eb0829c2f94d33
Signed-off-by: Chanho Park <chanho61.park@samsung.com>