workaround: drm: exynos: fix the status of drm fimd after resume
When FIMD is resumed, it is initialized again,
not restored to the status of that before suspended.
Therefore there must be no window which is enabled already,
because every window must be initialized again.

With this patch, page flip ioctl is not triggered
automatically, causing touching is needed to turn on
the display.

Change-Id: If1878ebdaadc3d3d56e3189af3e8c02c445c6293
Signed-off-by: Hyungwon Hwang <human.hwang@samsung.com>