workaround: drm: exynos: fix the status of drm fimd after resume
When FIMD is resumed, it is initialized again,
not restored to the status of that before suspended.
Therefore there must be no window which is enabled already,
because every window must be initialized again.

With this patch, page flip ioctl is not triggered
automatically, causing touching is needed to turn on
the display.

Change-Id: Ifcf94947d835ec9f4dbbef22949c6ae69705cda5
Signed-off-by: Hyungwon Hwang <human.hwang@samsung.com>