dma: pl330: Implement the dma channel pause/resume interface
This patch implements the dma pause and resume function for the
pl330 dmac driver. When the dmaengine_pause callback is called
from audio platform driver, the pl330 driver should handle the
pause command. And when the dmaengine_resume callback function
is called, the pl330 driver should handle the resume command.
Currently there is only the dmac channel state is set for pause
and resume function call, and there might have some continuous
work for the pl330 channel configuration in the future. However,
the odroidx2/u3 board can pause/release with this implementation.

Change-Id: I584106a4d7861fde308997e6843088fbf0a9b1bb
Signed-off-by: Huang Chao <chao7.huang@samsung.com>