ARM: dts: exynos4412-odroid-common: disable 'always on' for BUCK8 regulator
On Odroid U2/U3 BUCK8 is used for providing power to also to P3V3
source, which is also connected to LAN9730 chip's nRESET signal. To
reset lan chip on system reboot, the BUCK8 output should not be used in
'always on' mode. This change has no impact on X/X2 boards.

Signed-off-by: Kamil Debski <k.debski@samsung.com>
Change-Id: I27f942711c53ae9abf5a71034388d1f0081aa359
Signed-off-by: Marek Szyprowski <m.szyprowski@samsung.com>