Send process status in SCM_PROCINFO
This introduces a new SCM type called SCM_PROCINFO to allow the direct
attaching of process status to SCM, which is significantly more
efficient and will reliably avoid the race with the round-trip over
procfs. This is optional and can be turned on by setting SO_PASSPROC.

To achieve that, new struct called unix_skb_parms_scm had to be created,
because otherwise unix_skb_parms would be too big.

This patch is reworked version of:
https://lkml.org/lkml/2014/1/13/40

Change-Id: Idb488f713a9382c217f8c7f4e358d7138ecbb8c9
Signed-off-by: Jan Kaluza <jkaluza@redhat.com>
Signed-off-by: Piotr Wilczek <p.wilczek@samsung.com>