ASoC: max98090: Add jack irq status flag
The max98090 interrupt handler currently read the device status
register and will make real action only when the JDET bit is set.
But since this bit is not set when jack status has really changed,
we can use the jack status register to check the current jack status,
and use this jack irq status flag to record the current jack status.

Change-Id: I6bd5ebe4104b2076860660f1fee409a4eae1b6df
Signed-off-by: Huang Chao <chao7.huang@samsung.com>