ASoC: max98090: Refactor max98090_interrupt handler for jack detection
The max98090_interrupt handler currently will read the device status
register. After that, it will really acts only when the JDET bit is set.
However, when max98090_interrupt handler is triggered and jack status
register has really changed, but the JDET bit of device status register
is not set for unknown reason, so the jack status cannot be reported.

This patch adds a workaround solution for jack detection on odroidu3
board, by manually setting the JDET bit which indicates that the jack
status has just changed instantly.

Change-Id: I48ff1abb8a6553fab656c50fe1054bac8f398a4e
Signed-off-by: Huang Chao <chao7.huang@samsung.com>