ASoC: max98090: Refactor max98090_interrupt handler for jack detection
When max98090_interrupt handler is triggered, the jack status register
has really changed, but the JDET bit of the device status register is
not set for some reason, so the jack status cannot be reported.

This patch can provide a workaround solution for max98090 interrupt jack
detection on odroid board, by checking the jack status register which is
changed instantly when jack status changed.

Change-Id: I48ff1abb8a6553fab656c50fe1054bac8f398a4e
Signed-off-by: Huang Chao <chao7.huang@samsung.com>