clk: samsung: add new variable to compensate the gap for control register
PLL35xx and PLL2550 are very similar, and so they share the code for
their control. PLL2550 in Exynos5422 have different offsets to control
register from the base of them. mainline linux have the variable for it.
But because of too much difference between the mainline linux and
Tizen 3.10. It cannot be adopted with few commits. So I made this patch
to accomplish this purpose with small changes.

Change-Id: Iae2192a62d9d9cc6fcd24fb009c7a41f89acb8f3
Signed-off-by: Hyungwon Hwang <human.hwang@samsung.com>