exynos5420: add clock driver for exynos5420
This patch adds the clock driver of exynos5420. This driver is almost
same with the one in mainline linux kernel. There are too many commits
to this driver and the gap between exynos clock framework in Tizen 3.10
and Linux mainline 3.18 makes it hard and meaningless(in my opinion) to
get the original patches. So I modified only the interface between this
and the core of exynos clock framework to make it adoptable in Tizen
kernel 3.10.

Change-Id: I566d7b1fcbfcf84427276ac5246c4d81f094f0e7
Signed-off-by: Hyungwon Hwang <human.hwang@samsung.com>