arm: dts: exynos4412: add cpu node
This patch adds cpu node for Exynos4412. Because headsmp.S distinguish
a core by using the combination of cluster id and core id after the
commit 'add support secondary core bootup in big.LITTLE processor', but
when dt node for cpu is not provided, the core ids are assigned from 0
with the cluster id 0. But Exynos4412's cluster id for its 4 cores is
0xa. So this patch is needed for the cores to get the proper mpidr.

Change-Id: I2248fc332a4fbd05dc1983fcfa58f312e454c622
Signed-off-by: Hyungwon Hwang <human.hwang@samsung.com>